package rename, test with 2.8.3, sonarlint
