+++ spring-security/README.md
@@ -56,7 +56,7 @@ Then it auto-configures beans, that are required to initialize the Spring Boot a
 
 Auto-configuration class | Description
 ---- | --------
-[HybridAuthorizationAutoConfiguration]() | Creates a converter ([XsuaaTokenAuthorizationConverter](/src/main/java/com/sap/cloud/security/token/authentication/XsuaaTokenAuthorizationConverter.java))that removes the xsuaa application identifier from the scope names to enable local scope checks using [Spring's common built-in expression](https://docs.spring.io/spring-security/site/docs/current/reference/html5/#el-common-built-in) `hasAuthority`.
+[HybridAuthorizationAutoConfiguration]() | Creates a converter ([XsuaaTokenAuthorizationConverter](/com/sap/cloud/spring/token/authentication/XsuaaTokenAuthorizationConverter.java))that removes the xsuaa application identifier from the scope names to enable local scope checks using [Spring's common built-in expression](https://docs.spring.io/spring-security/site/docs/current/reference/html5/#el-common-built-in) `hasAuthority`.
 [HybridIdentityServicesAutoConfiguration](HybridAuthorizationAutoConfiguration.java) | Configures a `JwtDecoder` which is able to decode and validate tokens from Xsuaa and Identity service. Furthermore it registers the `XsuaaServiceConfiguration` and `IdentityServiceConfiguration` classes, that gets configured with `xsuaa.*` and `identity.*` properties.
 
 You can gradually replace auto-configurations as explained [here](https://docs.spring.io/spring-boot/docs/current/reference/html/using-boot-auto-configuration.html).
