+++ spring-security/src/main/java/com/sap/cloud/security/spring/config/XsuaaServiceConfigurations.java
@@ -1,0 +1,22 @@
+package com.sap.cloud.security.spring.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.NestedConfigurationProperty;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@ConfigurationProperties("sap.security.services")
+public class XsuaaServiceConfigurations {
+    @NestedConfigurationProperty
+    private List<XsuaaServiceConfiguration> xsuaa = new ArrayList<>();
+
+    public List<XsuaaServiceConfiguration> getConfigurations() {
+        return this.xsuaa;
+    }
+
+    public void setXsuaa(List<XsuaaServiceConfiguration> xsuaa) {
+        this.xsuaa = xsuaa;
+    }
+
+}
