+++ spring-security/src/main/resources/META-INF/services/com.sap.cloud.security.token.TokenFactory
@@ -1,0 +1,1 @@
+com.sap.cloud.security.servlet.HybridTokenFactory2
