+++ spring-security/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,3 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.sap.cloud.security.autoconfig.HybridIdentityServicesAutoConfiguration,\
-com.sap.cloud.security.autoconfig.HybridAuthorizationAutoConfiguration
+com.sap.cloud.security.spring.autoconfig.HybridIdentityServicesAutoConfiguration,\
+com.sap.cloud.security.spring.autoconfig.HybridAuthorizationAutoConfiguration
